Interruption Trap its work
kernel/arch/i386/interrupt.c:
try to get the number of interrupt as parameter from assebly

kernel/arch/i386/trap.S:
add define for save common registers and restore it and push $intno
to interrupt_handler function as parameter

kernel/include/kernel/interrupt.h:
change definition of interrupt_handler to get the interruption num as
paramter

kernel/kernel/kernel.c:
add a while to keep the processor working and trap the interruptions
correctly.

TO IMPROVE:
when an interruption is recived the os die, is needed to fix it.

Signed-off-by: Josue David Hernandez Gutierrez <megadavidtor@gmail.com>